Refactor common functionality to base class for transforms
